consolidate to one modal source for CDP
